Windows port: experimental deepseek_prefill: fix narrowing conversions
- attn_res_gather_softmax: static_cast CoreCoord x/y (size_t) to
  uint32_t in SetRuntimeArgs vector
- combine, dispatch: static_cast packet_header_size_bytes (size_t) to
  uint32_t in CBFormatDescriptor::page_size designated initializer;
  static_cast MeshCoordinateRange/MeshShape accessors (size_t) to
  uint32_t in runtime-args initializer lists
- zero_padded_kv_cache: static_cast CB index (uint32_t) to uint8_t in
  CBFormatDescriptor::buffer_index designated initializer

Note: attn_res_gather_softmax, attn_res_weighted_reduce_nc,
inbound_socket_service_sync, outbound_socket_service_sync,
pack_scaled_fp8_kv_cache, per_token_cast_to_fp8,
rotary_embedding_indexed, update_padded_kv_cache, and
zero_padded_kv_cache still fail to build in this tree: their
generated .obj/.lib output paths exceed the 260-char MAX_PATH limit
under .build/windows-msvc, causing C1083/LNK1104 regardless of source
correctness. This needs a build-tree-location fix, not a source fix.

+++ ttnn/cpp/ttnn/operations/experimental/deepseek_prefill/attn_res_gather_softmax/device/attn_res_gather_softmax_program_factory.cpp
@@ -431,8 +431,8 @@ AttnResGatherSoftmaxMeshWorkloadFactory::cached_program_t AttnResGatherSoftmaxMe
              tiles,
              start_tile,
              my_rank,
-             gather_physical.x,
-             gather_physical.y,
+             static_cast<uint32_t>(gather_physical.x),
+             static_cast<uint32_t>(gather_physical.y),
              total.buffer()->address()});
 
         tt::tt_metal::SetRuntimeArgs(program, compute_kernel_id, core, {rows, tiles, start_tile});
